Final API Audit changes
