+++ src/layers/core/grid-cell-layer/grid-cell-layer-fragment.glsl
