+++ src/layers/core/grid-cell-layer/grid-cell-layer-vertex-64.glsl
