+++ src/layers/core/grid-cell-layer/grid-cell-layer-vertex.glsl
