+++ src/layers/core/grid-layer/grid-aggregator.js
