+++ src/layers/core/hexagon-cell-layer/hexagon-cell-layer-fragment.glsl
