+++ src/layers/core/hexagon-cell-layer/hexagon-cell-layer-vertex-64.glsl
