+++ src/layers/core/hexagon-layer/hexagon-aggregator.js
